* commit 'b536822080e281b74a7d1409213188829e25fcf8':
  Update posda python lib, and add BackgroundHelloWorldWithInput.py

+++ posda/posdatools/python/posda/background/process.py
@@ -3,12 +3,15 @@ from .email import send_email
 from ..main import insert_file
 from ..main.downloadablefile import make_csv
 from ..queries import Query
+from ..util import printe
 
 import sys
 import os
 from io import StringIO
 import tempfile
 from datetime import datetime
+
+class MissingActivityIdError(KeyError): pass
 
 class BackgroundProcess:
     """Represents a background process
@@ -17,14 +20,16 @@ class BackgroundProcess:
     as well as preparing reports and email.
     """
 
-    def __init__(self, invoc_id, notify_address):
+    def __init__(self, invoc_id, notify_address, activity_id=None):
         self.invoc_id = invoc_id
         self.notify_address = notify_address
+        self.activity_id = activity_id
         self.reports = {}
 
         self.parent_pid = None
         self.child_pid = None
         self.input_line_query = None
+        self.input_line_count = 0
 
         for row in Query("CreateBackgroundSubprocess").run(
                 subprocess_invocation_id=invoc_id,
@@ -32,6 +37,10 @@ class BackgroundProcess:
                 foreground_pid=os.getpid(),
                 user_to_notify=notify_address):
             self.background_id = row.background_subprocess_id
+
+        if activity_id is not None:
+            Query('InsertActivityTaskStatus').execute(activity_id, invoc_id)
+
 
         self._log_args()
 
@@ -52,7 +61,14 @@ class BackgroundProcess:
         """For when good things just need to turn bad"""
         (self.parent_pid, self.child_pid) = real_daemonize()
 
-        # redirect stdout to a stringio, for email later
+
+        # turn email into a real report
+        email_report = self.create_report('Email')
+        print(self.email.getvalue(), file=email_report)
+        self.email = email_report
+
+
+        # redirect stdout to email
         self.real_stdout = sys.stdout
         sys.stdout = self.email
         self.log_input_count(self.input_line_count)
@@ -98,12 +114,32 @@ class BackgroundProcess:
         self.reports[name] = tempfile.NamedTemporaryFile(mode="w", delete=False)
         return self.reports[name]
 
-    def finish(self):
+
+    def _finish_activity_task_status(self, final_status_message: str=None) -> None:
+        if self.activity_id is None:
+            return
+
+        query = "FinishActivityTaskStatus"
+        if final_status_message == "Schedule Complete - Manual Process Follows":
+            query = "UpdateActivityTaskStatusForManualUpdate"
+
+        if final_status_message is None:
+            final_status_message = "Complete - no status specified"
+
+        Query(query).execute(
+            final_status_message,
+            self.activity_id,
+            self.invoc_id
+        )
+
+
+    def finish(self, final_status_message: str=None) -> None:
         """Indicate that the BackgroundProcess has finished.
 
         Log the completion time to the database, close all
         reports, load them into Posda and print the API URL to the email,
-        then send the email.
+        then send the email. Update Activity Task Status if we were
+        created with an activity_id.
         """
         Query("AddCompletionTimeToBackgroundProcess").execute(
             background_subprocess_id=self.background_id
@@ -112,14 +148,57 @@ class BackgroundProcess:
         print("Background process ended at:", self.finish_time)
         print("Total time elapsed:", self.finish_time - self.start_time)
 
+        self._finish_activity_task_status(final_status_message)
+
         for report_name, report in self.reports.items():
+            # skip the email for now, because if we close it we can't write
+            # any further lines
+            if report_name == 'Email':
+                continue
             report.close()
             file_id = insert_file(report.name)
             os.unlink(report.name)
             url = make_csv(file_id)
+            Query('CreateBackgroundReport').execute(
+                background_subprocess_id=self.background_id,
+                file_id=file_id,
+                name=report_name
+            )
             print(f"Report '{report_name}': {url}")
 
-        send_email(self.notify_address,
-                   "Posda job complete",
-                   self.email.getvalue())
+        # close email last
+        self.email.close()
+        email_file_id = insert_file(self.email.name)
+        os.unlink(self.email.name)
 
+        for report_id, in Query('CreateBackgroundReport').run(
+            background_subprocess_id=self.background_id,
+            file_id=email_file_id,
+            name='Email'
+        ):
+            send_email('nobody', self.notify_address, report_id, 'posda.background.process')
+        # TODO: make this 'send email' via posda
+        # send_email(self.notify_address,
+        #            "Posda job complete",
+        #            self.email.getvalue())
+
+
+    def set_activity_status(self, status: str, time_remaining: str=None) -> None:
+        if self.activity_id is None:
+            raise MissingActivityIdError("This BackgroundProcess was created "
+                                         "without an activity_id, so Activity "
+                                         "Status cannot be set.")
+
+        if time_remaining is not None:
+            Query("UpdateActivityTaskStatusAndCompletionTime").execute(
+                status_text=status,
+                expected_completion_time=time_remaining,
+                activity_id=self.activity_id,
+                subprocess_invocation_id=self.invoc_id
+            )
+        else:
+            Query("UpdateActivityTaskStatus").execute(
+                status_text=status,
+                activity_id=self.activity_id,
+                subprocess_invocation_id=self.invoc_id
+            )
